+++ crates/ika-core/src/dwallet_mpc/tests/mod.rs
@@ -1,1 +1,2 @@
 mod network_dkg;
+mod utils;
